Respond to PR feedback
As title.
